+++ types/mqmho.d.ts
@@ -1,9 +1,7 @@
 declare module "ibmmq" {
   /**
    * This is a class containing the fields needed for the MQCMHO
-   * (MQ Create Message Handle Options) structure. See the
-   * {@link https://www.ibm.com/support/knowledgecenter/en/SSFKSJ_latest/com.ibm.mq.ref.dev.doc/q095320_.htm |MQ Knowledge Center}
-   * for more details on the usage of each field.
+   * (MQ Create Message Handle Options) structure. 
    */
   class MQCMHO {
     Options: number |MQC_MQCMHO[];    
@@ -11,9 +9,7 @@ declare module "ibmmq" {
 
   /**
    * This is a class containing the fields needed for the MQDMHO
-   * (MQ Delete Message Handle Options) structure. See the
-   * {@link https://www.ibm.com/support/knowledgecenter/en/SSFKSJ_latest/com.ibm.mq.ref.dev.doc/q096320_.htm |MQ Knowledge Center}
-   * for more details on the usage of each field.
+   * (MQ Delete Message Handle Options) structure. 
    */
   class MQDMHO {
     Options: number|MQC_MQDMHO[];