+++ types/mqrfh2.d.ts
@@ -10,9 +10,7 @@ declare module "ibmmq" {
    * @class
    * @classdesc
    * This is a class containing the fields needed for the
-   * MQ RFH2 Header structure. See the
-   * {@link https://www.ibm.com/support/knowledgecenter/en/SSFKSJ_latest/com.ibm.mq.ref.dev.doc/q096110_.htm|MQ Knowledge Center}
-   * for more details on the usage of each field.
+   * MQ RFH2 Header structure. 
    * Not all of the underlying fields may be exposed in this object.
    */
   class MQRFH2 {